Add EIP210 contract to provide blockhashes for arbitrary block numbers.
Currently smart contracts can access block hashes for the previous 256 blocks, but no further. Effective "purchase-then-reveal" implementation requires that block hashes remain available indefinitely; see "Blockhash-Minus-256" problem. EIP-210 is a proposed rules change that makes blockhashes available from any previous block.

The EIP210 contract simulates this functionality. When a blockhash is requested the result is permanently stored in the contract for consistency. Depending on when it is requested, the hash may be real (for blocks at -256 to -1) or simulated. Simulated hashes are based on the last bits of the requested block and the first bits of the current block; this simulated hash will only change every 256 blocks, making it time-consuming for an attacker game. Further, if any other transaction causes the non-reverted read of the that block's hash, the result will be stored and invariate for the attacker.

Recommended usage is thus: 1. request block hashes as required by the transaction being processed, and 2. find opportunities to request the hashes of other significant blocks, if possible; i.e. the blocks which an attacker might attempt to game.

For example in a purchase-then-reveal application, where a pseudorandom number is generated from a blockhash that occurred between the two steps, the contract should store block numbers that will eventually be revealed and store the computed-or-simulated hashes of those blocks as a side-effect of transactions unrelated to the reveal operation.