RaraCollectibleEmitter: significant design rework. Use daily "activation" model.
In a given mining period (arbitrary length but assumed daily for clarity), only "activated" tokens produce mining rewards. Multiple mining "pools" may be defined, each with a particular staking token, (optionally) a token type, and a mining power reward. Activation of a token is automatic the first time that particular token is deposited, but after that must be manually performed daily. Each pool defines the cost of activation as a quantity of ERC721 tokens burned, which (like the staking token) may optionally be a particular type.

Like activation, mining is performed daily. The accumulated reward from the emitter is divided among the activated mining power for that day (pools do not have "weight" or "score" as they do in LP mining; a "pool" serves only to organize a particular mining token and its power, activation requirements, etc.

Testing is fairly comprehensive but `updatePeriod` requires more extensive testing when dealing with changes in `initTime` and `periodAnchor` s.t. that they do not flow immediately from one to the next.