Develop (#1)
* add RaraVotingToken, BufferTokenEmitter, and tests

* update migration

* update migration: deploy router later, update Rara decimals

* include dex router in migration

* include flattened code for testnet deployed contracts

* update "init code hash" for uniswap pair

* update Rara decmials: 8 -> 18

* update mining pool: add setUnlockBlock to adjust unlock time (up until it actually happens)

* update VotingMembershipListener to use ERC165 instead of `isVotingMembershipListener`

* WIP "RaraCollectible" token (zebra, shiba, etc.)

* migration fixes

* token fixes

* fix token tests

* RaraMiningPool: test setUnlockBlock

* include build/contracts

* finish RaraCollectible draft

* add Dex Mocker migration; update contract build files with new contracts and deployments

* more testing for RaraCollectible

* Add RaraCollectibleMining, a contract for mining Rara based on a token collection and/or voting registry level.

* WIP RaraCollectibleMining tests

* include contract build updates

* fixes to RaraCollectibleMining

* Add WithdrawFeeValidatingStakeManager and tests

* Validating StakeManager: adds tests and migration; updates builds

* some collectible token fixes

* add sales / blind box / resale market / token conversion contracts

* fix converter and add basic tests

* add Jamie Chua set 1 migration

* blind box: use timestamps, not blocks

* add tests for converter

* Add tests (and fixes) to RaraCollectibleMining

* Refactor blind sale contract structure and naming schemes

* add tests for RaraCollectibleMining

* Add EIP210 contract to provide blockhashes for arbitrary block numbers.

Currently smart contracts can access block hashes for the previous 256 blocks, but no further. Effective "purchase-then-reveal" implementation requires that block hashes remain available indefinitely; see "Blockhash-Minus-256" problem. EIP-210 is a proposed rules change that makes blockhashes available from any previous block.

The EIP210 contract simulates this functionality. When a blockhash is requested the result is permanently stored in the contract for consistency. Depending on when it is requested, the hash may be real (for blocks at -256 to -1) or simulated. Simulated hashes are based on the last bits of the requested block and the first bits of the current block; this simulated hash will only change every 256 blocks, making it time-consuming for an attacker game. Further, if any other transaction causes the non-reverted read of the that block's hash, the result will be stored and invariate for the attacker.

Recommended usage is thus: 1. request block hashes as required by the transaction being processed, and 2. find opportunities to request the hashes of other significant blocks, if possible; i.e. the blocks which an attacker might attempt to game.

For example in a purchase-then-reveal application, where a pseudorandom number is generated from a blockhash that occurred between the two steps, the contract should store block numbers that will eventually be revealed and store the computed-or-simulated hashes of those blocks as a side-effect of transactions unrelated to the reveal operation.

* Refactor conversion; adds BankableCollectibleConverter.

Converters now emit a Conversion event containing both the tokenIds and tokenTypes input to the recipe.

A Bankable converter allows users to "bank" (burn) tokens for later use in conversion; this allows the `convert` step to use fewer tokens. The main purpose is to split a very large conversion into multiple steps. At the extreme, any recipe can be used for conversion with no tokens input at all, provided the necessary tokens have already been banked.

* Refactor blind sales; add GachaRack, Revealable interfaces, implementation, tests.

Slightly revamps the IBlindSalePrizeBag, moving `availableSupply` to the base IBlindSale. Adds `IBlindSaleGachaRack` and `IBlindSaleRevealable` as subinterfaces.

IBlindSaleGachaRack: allows multiple games with users defaulting to a particular gameId; each game has its own prizes and probabilities. Two possible uses: support 2-step "purchase then reveal" structure while also supporting probability updates; specify the probability / prizes as a new _game_, rather than a modification of an existing game, and already purchased draws can keep the old structure.

IBlindSaleRevealable: draws are purchased in an "unrevealed" state, and can be manually revealed later. Allows 2-step purchase-then-reveal operation s.t. an attacker cannot revert the purchase transaction if they don't like the result.

Also adds base implementations of multiple interfaces and a preset that combines them.

* Update Classifieds "TradeStatusChange" event.

* Add a collection of "Animal" contracts (deployed to testnet).

* update migrations through RaraAnimal deployment

* updates to build/contracts/*

* Fixes a drawId indexing problem of GachaRack draws (and adds regression test).

* add FlowSupplementEmitter and tests

Acts as an intermediate emitter (relying on an upstream emitter as token source) that supplements the flow with tokens from its own supply, equal in number to those claimed from the upstream emitter.

* fixes to GachaRack (and tests)

* slight optimization to FlowSupplementEmitter

* RaraCollectibleEmitter: significant design rework. Use daily "activation" model.

In a given mining period (arbitrary length but assumed daily for clarity), only "activated" tokens produce mining rewards. Multiple mining "pools" may be defined, each with a particular staking token, (optionally) a token type, and a mining power reward. Activation of a token is automatic the first time that particular token is deposited, but after that must be manually performed daily. Each pool defines the cost of activation as a quantity of ERC721 tokens burned, which (like the staking token) may optionally be a particular type.

Like activation, mining is performed daily. The accumulated reward from the emitter is divided among the activated mining power for that day (pools do not have "weight" or "score" as they do in LP mining; a "pool" serves only to organize a particular mining token and its power, activation requirements, etc.

Testing is fairly comprehensive but `updatePeriod` requires more extensive testing when dealing with changes in `initTime` and `periodAnchor` s.t. that they do not flow immediately from one to the next.

* rename gacha rack preset as "...ManagedFlowGachaRack" to allow other types of management

* Add BaseBlindCollectibleGachaRackRegularResupply and tests.

An alternative to the LimitedFlow model, this type of GachaGame automatically refreshes its supply up to a set level at the end of every sale period (determined by a duration and an anchor time).

* update nft_animals deployment for the new GachaRack game

* include in previous

* WIP: direct collectible sales with regular resupply and snap-up pricing.

TODO items:

1. Fill in the FractionalExponentsLibrary
2. Add unit testing
3. Add a migration w/ settings for animal tokens

* additional testing regarding collectible mining period changes

* fix previous

* include "claimProceeds" testing for sale contracts

* direct collectible sales (resupply and snap-up pricing), with testing.

* Update direct sale contracts

* Classifieds: allow management (canceling of trades) by admin

* WIP: revised GachaRack

* fix EIP210 test

* WIP gacha rack

* update tests

* update gacha rack and EIP210

* add testing for ManagedResupplyGachaRack

* add ERC721CollectibleReferenceOracle.

The oracle reports tokenTypes for ERC721 tokens that don't implement our bespoke "Collectible" extension.

* Add contracts to deal with Binance-isssued Animal tokens

* Adds Binance proxy exchanger, migrations